delete account service

+++ src/main/java/com/example/jobcrony/controller/UserController.java
@@ -2,6 +2,7 @@ package com.example.jobcrony.controller;
 
 import com.example.jobcrony.dtos.requests.UpdatePasswordRequest;
 import com.example.jobcrony.dtos.responses.GenericResponse;
+import com.example.jobcrony.exceptions.InvalidConfirmationException;
 import com.example.jobcrony.exceptions.SendMailException;
 import com.example.jobcrony.exceptions.UserNotFoundException;
 import com.example.jobcrony.exceptions.WrongPasswordException;
@@ -27,8 +28,8 @@ public class UserController {
         return userService.forgotPassword(email);
     }
 
-    @DeleteMapping("/deleteAccount")
-    public ResponseEntity<GenericResponse<String>> deleteAccount(){
-        return ResponseEntity.ok().body(userService.deleteAccount());
+    @DeleteMapping("/deleteAccount/{confirmationKeyword}")
+    public ResponseEntity<GenericResponse<String>> deleteAccount(@PathVariable String confirmationKeyword) throws UserNotFoundException, InvalidConfirmationException {
+        return ResponseEntity.ok().body(userService.deleteAccount(confirmationKeyword));
     }
 }
